chore: move stake page into new stake component
